refactor: Rename for clarity
